+++ packages/aio-commerce-lib-admin-ui/source/order-view-buttons/index.ts
@@ -21,6 +21,7 @@
  * @packageDocumentation
  */
 
+export { getOrderViewButtonAclResourceId } from "./acl-resource-id";
 export {
   okOrderViewButtonResponse,
   orderViewButtonErrorResponse,
