+++ packages/aio-commerce-lib-admin-ui/test/unit/api/lib/acl-resource-id.test.ts
@@ -12,13 +12,7 @@
 
 import { describe, expect, it } from "vitest";
 
-import {
-  getAclResourceId,
-  getGridColumnAclResourceId,
-  getMassActionAclResourceId,
-  getMenuAclResourceId,
-  getOrderViewButtonAclResourceId,
-} from "#api/lib/acl-resource-id";
+import { getAclResourceId } from "#api/lib/acl-resource-id";
 
 describe("getAclResourceId", () => {
   it("converts hyphens to underscores and lowercases (cross-repo contract fixture)", () => {
@@ -42,151 +36,5 @@ describe("getAclResourceId", () => {
   it("returns empty string for blank input", () => {
     expect(getAclResourceId("")).toBe("");
     expect(getAclResourceId("   ")).toBe("");
-  });
-});
-
-describe("getMenuAclResourceId", () => {
-  it("produces the menu leaf id (cross-repo contract fixture)", () => {
-    expect(
-      getMenuAclResourceId("approval-dashboard-app", "approval_dashboard"),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_approval_dashboard_app_menu_approval_dashboard",
-    );
-  });
-
-  it("sanitizes hyphens in both metadataId and menuId independently", () => {
-    expect(getMenuAclResourceId("My-App", "My-Item")).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_my_app_menu_my_item",
-    );
-  });
-
-  it("trims whitespace in both segments before sanitizing", () => {
-    expect(
-      getMenuAclResourceId(
-        "  approval-dashboard-app  ",
-        "  approval_dashboard  ",
-      ),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_approval_dashboard_app_menu_approval_dashboard",
-    );
-  });
-
-  it("produces the same id whether menuId uses hyphens or underscores", () => {
-    expect(
-      getMenuAclResourceId("approval-dashboard-app", "approval-dashboard"),
-    ).toBe(
-      getMenuAclResourceId("approval-dashboard-app", "approval_dashboard"),
-    );
-  });
-
-  it("returns empty string when metadataId is blank", () => {
-    expect(getMenuAclResourceId("", "approval_dashboard")).toBe("");
-    expect(getMenuAclResourceId("   ", "approval_dashboard")).toBe("");
-  });
-});
-
-describe("getGridColumnAclResourceId", () => {
-  it("produces the order grid-column leaf id (cross-repo contract fixture)", () => {
-    expect(
-      getGridColumnAclResourceId(
-        "approval-dashboard-app",
-        "order",
-        "order_status",
-      ),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_approval_dashboard_app_order_gridcolumns_order_status",
-    );
-  });
-
-  it("supports product and customer entities", () => {
-    expect(
-      getGridColumnAclResourceId("acme-promotions", "product", "approval-req"),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_acme_promotions_product_gridcolumns_approval_req",
-    );
-    expect(
-      getGridColumnAclResourceId("acme-promotions", "customer", "Tier"),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_acme_promotions_customer_gridcolumns_tier",
-    );
-  });
-
-  it("trims and sanitizes both metadataId and columnId", () => {
-    expect(
-      getGridColumnAclResourceId("  My-App  ", "order", "  My-Col  "),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_my_app_order_gridcolumns_my_col",
-    );
-  });
-
-  it("returns empty string when metadataId is blank", () => {
-    expect(getGridColumnAclResourceId("", "order", "order_status")).toBe("");
-    expect(getGridColumnAclResourceId("   ", "order", "order_status")).toBe("");
-  });
-});
-
-describe("getMassActionAclResourceId", () => {
-  it("produces the order mass-action leaf id (cross-repo contract fixture)", () => {
-    expect(
-      getMassActionAclResourceId(
-        "approval-dashboard-app",
-        "order",
-        "bulk-approve",
-      ),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_approval_dashboard_app_order_massactions_bulk_approve",
-    );
-  });
-
-  it("supports product and customer entities", () => {
-    expect(
-      getMassActionAclResourceId("acme-promotions", "product", "flag-review"),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_acme_promotions_product_massactions_flag_review",
-    );
-    expect(
-      getMassActionAclResourceId("acme-promotions", "customer", "assign-tier"),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_acme_promotions_customer_massactions_assign_tier",
-    );
-  });
-
-  it("trims and sanitizes both metadataId and actionId", () => {
-    expect(
-      getMassActionAclResourceId("  My-App  ", "order", "  Bulk Approve  "),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_my_app_order_massactions_bulk_approve",
-    );
-  });
-
-  it("returns empty string when metadataId is blank", () => {
-    expect(getMassActionAclResourceId("", "order", "bulk-approve")).toBe("");
-    expect(getMassActionAclResourceId("   ", "order", "bulk-approve")).toBe("");
-  });
-});
-
-describe("getOrderViewButtonAclResourceId", () => {
-  it("produces the order view-button leaf id (cross-repo contract fixture)", () => {
-    expect(
-      getOrderViewButtonAclResourceId(
-        "approval-dashboard-app",
-        "approve-order",
-      ),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_approval_dashboard_app_order_viewbuttons_approve_order",
-    );
-  });
-
-  it("trims and sanitizes both segments", () => {
-    expect(
-      getOrderViewButtonAclResourceId("  My-App  ", "  Approve Order  "),
-    ).toBe(
-      "Magento_CommerceBackendUix::adminuisdk_app_my_app_order_viewbuttons_approve_order",
-    );
-  });
-
-  it("returns empty string when metadataId is blank", () => {
-    expect(getOrderViewButtonAclResourceId("", "approve-order")).toBe("");
-    expect(getOrderViewButtonAclResourceId("   ", "approve-order")).toBe("");
   });
 });
